Updated the pysb-units version in environment.yml. Moved the creation of model temp dir and loading of util and widget modules from the streamlit_app to the home page file.
